learn: more in depth testing best practices

+++ src/_first/StorageFactory.sol
@@ -10,7 +10,7 @@ contract StorageFactory {
         storages.push(new Storage());
     }
 
-    function sStore(uint _index, uint256 _number) public {
+    function sStore(uint256 _index, uint256 _number) public {
         storages[_index].store(_number);
     }
 }
